Added Loop
The program will now execute on a loop until the user prompts the
program to exit.
The program now uses switch statements instead of if, else if, else etc.
The program now uses BufferedReader instead of Scanner. I was having
problems with the while loop where after executing the addition
operation it would run a second time without any input from the user.
Switching to the BufferedReader fixed this.