Added Subtraction Functionality
Added the ability to do subtraction. Also re worked the way the program
gets the numbers for the operation. Now there is a method called
GetNumber that will prompt the user to enter a number and then returns
it.  These numbers are then passed into the addition or subtraction
methods and the result is returned.